edited files to attempt removal of duplicates creaeted during initial creation
