--- app/public/Input htmls/create.html
@@ -1,122 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Vision Finance</title>
-    <link rel="stylesheet" type="text/css" href="../css/inputs.css">
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.0/css/bootstrap.min.css" integrity="sha384-9aIt2nRpC12Uk9gS9baDl411NQApFmC26EwAOH8WgZl5MYYxFfc+NcPb1dKGj7Sk" crossorigin="anonymous">
-    
-  </head>
-<body>
-
-
-  
-
-  <form>
-
-    <topping>Create Account</topping>
-    <br>
-    <br>
-    <br>
-
-    <div class="form-row">
-      
-      
-
-        <!--------------------------------NAME/EMAIL--------------------------------->
-        <div class="form-group col-md-6">
-        <label for="inputFirstName">First Name</label>
-        <input type="first-name" class="form-control" id="inputFirstName" placeholder="John">
-      </div>
-      <div class="form-group col-md-6">
-        <label for="inputLastName">Last Name</label>
-        <input type="last-name" class="form-control" id="inputLastName" placeholder="Smith">
-      </div>
-    </div>
-
-    <div class="form-group">
-      <label for="inputEmail">Email</label>
-      <input type="text" class="form-control" id="inputEmail" placeholder="Jsmith@gmail.com">
-    </div>
-
-          <!--------------------------------PASSWORD--------------------------------->
-    <div class="form-row">
-      
-      <div class="form-group col-md-6">
-        
-        <label for="inputPassword">Password</label>
-        <input type="first-name" class="form-control" id="inputPassword" placeholder="Must be 8-20 characters long">
-      </div>
-      <div class="form-group col-md-6">
-        <label for="inputRtPassword">Re-Type Password</label>
-        <input type="last-name" class="form-control" id="inputRtPassword" placeholder="Re-Type Password">
-      </div>
-    </div>
-
-          <!--------------------------------SIGN IN/BACK--------------------------------->
-
-  
-
-    <br>
-    <button type="submit" class="btn btn-primary">Sign in</button>
-    <br>
-    <br>
-    <button type="back" class="btn" id= "back">Back</button>
-
-
-
-
-          <!--------------------------------LOCATION--------------------------------->
-
-        <!-- <div class="form-row">
-      <div class="form-group col-md-6">
-        <label for="inputCity">City</label>
-        <input type="text" class="form-control" id="inputCity">
-      </div>
-      <div class="form-group col-md-4">
-        <label for="inputState">State</label>
-        <select id="inputState" class="form-control">
-          <option selected>Choose...</option>
-          <option>...</option>
-        </select>
-      </div>
-      <div class="form-group col-md-2">
-        <label for="inputZip">Zip</label>
-        <input type="text" class="form-control" id="inputZip">
-      </div>
-    </div> -->
-
-
-
-      <!-- <div class="form-group">
-      <div class="form-check">
-        <input class="form-check-input" type="checkbox" id="gridCheck">
-        <label class="form-check-label" for="gridCheck">
-          Check me out
-        </label>
-      </div>
-    </div> -->
-
-
-
-
-
-
-  </form>
-   
-  
-
-
-
-
-
-
-
-
-
-
-
-
-</body>
-</html>
