--- app/public/Input htmls/personal.html
@@ -1,116 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Vision Finance</title>
-    <link rel="stylesheet" type="text/css" href="../css/inputs.css">
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.0/css/bootstrap.min.css"
-        integrity="sha384-9aIt2nRpC12Uk9gS9baDl411NQApFmC26EwAOH8WgZl5MYYxFfc+NcPb1dKGj7Sk" crossorigin="anonymous">
-
-</head>
-
-<body>
-
-
-
-
-    <form>
-
-        <topping>Personal Expenses</topping>
-        <br>
-        <br>
-        <br>
-
-
-        <div class="form-group row">
-            <label for="colFormLabelLg" class="col-sm-4 col-form-label col-form-label">Clothing</label>
-            <div class="col-sm-4">
-                <input type="email" class="form-control form-control" id="colFormLabelSm" placeholder="Planned">
-            </div>
-            <div class="col-sm-4">
-                <input type="email" class="form-control form-control" id="colFormLabelSm" placeholder="Received">
-            </div>
-        </div>
-
-        <div class="form-group row">
-            <label for="colFormLabelLg" class="col-sm-4 col-form-label col-form-label">Phone</label>
-            <div class="col-sm-4">
-                <input type="email" class="form-control form-control" id="colFormLabelSm" placeholder="Planned">
-            </div>
-            <div class="col-sm-4">
-                <input type="email" class="form-control form-control" id="colFormLabelSm" placeholder="Received">
-            </div>
-        </div>
-
-        <div class="form-group row">
-            <label for="colFormLabelLg" class="col-sm-4 col-form-label col-form-label">Personal Hygiene</label>
-            <div class="col-sm-4">
-                <input type="email" class="form-control form-control" id="colFormLabelSm" placeholder="Planned">
-            </div>
-            <div class="col-sm-4">
-                <input type="email" class="form-control form-control" id="colFormLabelSm" placeholder="Received">
-            </div>
-        </div>
-        <div class="form-group row">
-            <label for="colFormLabelLg" class="col-sm-4 col-form-label col-form-label">Subscriptions</label>
-            <div class="col-sm-4">
-                <input type="email" class="form-control form-control" id="colFormLabelSm" placeholder="Planned">
-            </div>
-            <div class="col-sm-4">
-                <input type="email" class="form-control form-control" id="colFormLabelSm" placeholder="Received">
-            </div>
-        </div>
-        <div class="form-group row">
-            <label for="colFormLabelLg" class="col-sm-4 col-form-label col-form-label">Recreation</label>
-            <div class="col-sm-4">
-                <input type="email" class="form-control form-control" id="colFormLabelSm" placeholder="Planned">
-            </div>
-            <div class="col-sm-4">
-                <input type="email" class="form-control form-control" id="colFormLabelSm" placeholder="Received">
-            </div>
-        </div>
-
-
-
-
-
-        <!------------------------------ADDING LABEL----------------------->
-
-
-
-
-        <div class="form-group row">
-            <button onclick=addItem() id="add-label" for="colFormLabelLg"
-                class="col-sm-4 col-form-label col-form-label btn">Add Label</button>
-        </div>
-
-
-
-
-
-
-
-        <!--------------------------------SIGN IN/BACK--------------------------------->
-
-
-        <br>
-        <br>
-        <button type="submit" class="btn btn-primary" id="continue">Continue</button>
-        <br>
-        <br>
-        <button type="back" class="btn" id="back">Back</button>
-
-
-
-
-
-
-    </form>
-
-    <script src="script.js"></script>
-
-</body>
-
-</html>
