+++ app/public/index.html
@@ -9,8 +9,8 @@
     <title>VisionFinance - finance your visions</title>
     <meta name="description" content="This is the landing page" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <!-- Favicon
-    <link rel="icon" type="image/x-icon" href="../assets/img/desk image.jpg" /> -->
+    <!-- Favicon-->
+    <link rel="icon" type="image/x-icon" href="/assets/img/favicon.ico" />
     <!-- Font Awesome icons (free version)-->
     <script src="https://use.fontawesome.com/releases/v5.13.0/js/all.js" crossorigin="anonymous"></script>
     <!-- Google fonts-->
@@ -19,7 +19,7 @@
     <!-- Third party plugin CSS-->
     <link href="https://cdnjs.cloudflare.com/ajax/libs/magnific-popup.js/1.1.0/magnific-popup.min.css" rel="stylesheet" />
     <!-- Core theme CSS (includes Bootstrap)-->
-    <link href="css/index-style.css" rel="stylesheet" />
+    <link href="/css/styles.css" rel="stylesheet" />
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous" />
 </head>
 
@@ -32,11 +32,9 @@
     <![endif]-->
 
     <nav class="navbar navbar-expand-lg navbar-light fixed-top py-3" id="mainNav">
-       
-       
-        <!-- <div class="container"> -->     
-
-<!-- 
+        <div class="container">
+            <a class="navbar-brand js-scroll-trigger" href="#page-top">Vision Finance</a
+        >
         <button
           class="navbar-toggler navbar-toggler-right"
           type="button"
@@ -44,15 +42,12 @@
           data-target="#navbarResponsive"
           aria-controls="navbarResponsive"
           aria-expanded="false"
-          aria-label="Toggle navigation">
-
-          
+          aria-label="Toggle navigation"
+        >
           <span class="navbar-toggler-icon"></span>
-        </button> -->
-
-
+        </button>
         <div class="collapse navbar-collapse" id="navbarResponsive">
-          <ul class="navbar-nav ml-auto my-2 my-lg-0 back">
+          <ul class="navbar-nav ml-auto my-2 my-lg-0">
             <li class="nav-item">
               <a class="nav-link js-scroll-trigger" href="#about">About</a>
             </li>
@@ -61,40 +56,38 @@
               >
             </li>
             <li class="nav-item">
-              <a class="nav-link js-scroll-trigger" href="#login">Login</a>
+              <a class="nav-link js-scroll-trigger" href="./login.html"
+                >Login</a
+              >
             </li>
             <li class="nav-item">
-                <a class="nav-link js-scroll-trigger" href="#signup">Signup</a>
+              <a class="nav-link js-scroll-trigger" href="./register.html"
+                >Signup</a
+              >
             </li>
-            </ul>
+          </ul>
         </div>
-        </div>
+      </div>
     </nav>
     <!-- Masthead-->
     <header class="masthead">
-        
-        <div class="container h-100">
-            
-            <div class="row h-100 align-items-center justify-content-center text-center">
-                
-                <div class="name text-uppercase font-weight-bold">
-                    <a>Vision Finance</a>
-                    
-                    
-                    
-                    
-                    
-                    <h1 class="text-uppercase text-white font-weight-bold">
-                        Let us help finance your vision
-                    </h1>
-                    <hr class="divider my-4" />
-                </div>
-                <div class="col-lg-8 align-self-baseline">
-                    <p class="text-blue font-weight-bold mb-5">
-                        Track your expense! Plan your budget! Visualize your dreams!
-                    </p>
-                    <a class="btn btn-primary btn-xl js-scroll-trigger" href="#about">Find Out More</a>
-
+      <div class="container h-100">
+        <div
+          class="row h-100 align-items-center justify-content-center text-center"
+        >
+          <div class="col-lg-10 align-self-end">
+            <h1 class="text-uppercase text-white font-weight-bold">
+              Let us help finance your vision
+            </h1>
+            <hr class="divider my-4" />
+          </div>
+          <div class="col-lg-8 align-self-baseline">
+            <p class="text-white-75 font-weight-light mb-5">
+              Track your expense! Plan your budget! Visualize your dreams!
+            </p>
+            <a class="btn btn-primary btn-xl js-scroll-trigger" href="#about"
+              >Find Out More</a
+            >
           </div>
         </div>
       </div>
@@ -124,7 +117,7 @@
         <h2 class="text-center mt-0">At Your Service</h2>
         <hr class="divider my-4" />
         <div class="row">
-          <div class="col-lg-4 col-md-6 text-center">
+          <div class="col-lg-3 col-md-6 text-center">
             <div class="mt-5">
               <i class="fas fa-4x fa-gem text-primary mb-4"></i>
               <h3 class="h4 mb-2">Create Budget</h3>
@@ -133,7 +126,7 @@
               </p>
             </div>
           </div>
-          <div class="col-lg-4 col-md-6 text-center">
+          <div class="col-lg-3 col-md-6 text-center">
             <div class="mt-5">
               <i class="fas fa-4x fa-laptop-code text-primary mb-4"></i>
               <h3 class="h4 mb-2">Track Expense</h3>
@@ -142,7 +135,7 @@
               </p>
             </div>
           </div>
-          <div class="col-lg-4 col-md-6 text-center">
+          <div class="col-lg-3 col-md-6 text-center">
             <div class="mt-5">
               <i class="fas fa-4x fa-globe text-primary mb-4"></i>
               <h3 class="h4 mb-2">Dashboard</h3>
@@ -155,11 +148,12 @@
       </div>
     </section>
     <!-- Footer-->
-    <footer >
-
-        A FOOTER
-      
-      
+    <footer class="bg-light py-5">
+      <div class="container">
+        <div class="small text-center text-muted">
+          Copyright © 2020 - Start Bootstrap
+        </div>
+      </div>
     </footer>
     <!-- Bootstrap core JS-->
     <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
