--- public/index.html
@@ -1,167 +1,0 @@
-<!DOCTYPE html>
-
-<html>
-<!--<![endif]-->
-
-<head>
-    <meta charset="utf-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <title>VisionFinance - finance your visions</title>
-    <meta name="description" content="This is the landing page" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <!-- Favicon-->
-    <link rel="icon" type="image/x-icon" href="/assets/img/favicon.ico" />
-    <!-- Font Awesome icons (free version)-->
-    <script src="https://use.fontawesome.com/releases/v5.13.0/js/all.js" crossorigin="anonymous"></script>
-    <!-- Google fonts-->
-    <link href="https://fonts.googleapis.com/css?family=Merriweather+Sans:400,700" rel="stylesheet" />
-    <link href="https://fonts.googleapis.com/css?family=Merriweather:400,300,300italic,400italic,700,700italic" rel="stylesheet" type="text/css" />
-    <!-- Third party plugin CSS-->
-    <link href="https://cdnjs.cloudflare.com/ajax/libs/magnific-popup.js/1.1.0/magnific-popup.min.css" rel="stylesheet" />
-    <!-- Core theme CSS (includes Bootstrap)-->
-    <link href="/css/styles.css" rel="stylesheet" />
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous" />
-</head>
-
-<body>
-    <!--[if lt IE 7]>
-      <p class="browsehappy">
-        You are using an <strong>outdated</strong> browser. Please
-        <a href="#">upgrade your browser</a> to improve your experience.
-      </p>
-    <![endif]-->
-
-    <nav class="navbar navbar-expand-lg navbar-light fixed-top py-3" id="mainNav">
-        <div class="container">
-            <a class="navbar-brand js-scroll-trigger" href="#page-top">Vision Finance</a
-        >
-        <button
-          class="navbar-toggler navbar-toggler-right"
-          type="button"
-          data-toggle="collapse"
-          data-target="#navbarResponsive"
-          aria-controls="navbarResponsive"
-          aria-expanded="false"
-          aria-label="Toggle navigation"
-        >
-          <span class="navbar-toggler-icon"></span>
-        </button>
-        <div class="collapse navbar-collapse" id="navbarResponsive">
-          <ul class="navbar-nav ml-auto my-2 my-lg-0">
-            <li class="nav-item">
-              <a class="nav-link js-scroll-trigger" href="#about">About</a>
-            </li>
-            <li class="nav-item">
-                <a class="nav-link js-scroll-trigger" href="#services">Services</a
-              >
-            </li>
-            <li class="nav-item">
-              <a class="nav-link js-scroll-trigger" href="./login.html"
-                >Login</a
-              >
-            </li>
-            <li class="nav-item">
-              <a class="nav-link js-scroll-trigger" href="./register.html"
-                >Signup</a
-              >
-            </li>
-          </ul>
-        </div>
-      </div>
-    </nav>
-    <!-- Masthead-->
-    <header class="masthead">
-      <div class="container h-100">
-        <div
-          class="row h-100 align-items-center justify-content-center text-center"
-        >
-          <div class="col-lg-10 align-self-end">
-            <h1 class="text-uppercase text-white font-weight-bold">
-              Let us help finance your vision
-            </h1>
-            <hr class="divider my-4" />
-          </div>
-          <div class="col-lg-8 align-self-baseline">
-            <p class="text-white-75 font-weight-light mb-5">
-              Track your expense! Plan your budget! Visualize your dreams!
-            </p>
-            <a class="btn btn-primary btn-xl js-scroll-trigger" href="#about"
-              >Find Out More</a
-            >
-          </div>
-        </div>
-      </div>
-    </header>
-    <!-- About-->
-    <section class="page-section bg-primary" id="about">
-      <div class="container">
-        <div class="row justify-content-center">
-          <div class="col-lg-8 text-center">
-            <h2 class="text-white mt-0">We've got what you need!</h2>
-            <hr class="divider light my-4" />
-            <p class="text-white-50 mb-4">
-              Start Bootstrap has everything you need to get your new website up
-              and running in no time! Choose one of our open source, free to
-              download, and easy to use themes! No strings attached!
-            </p>
-            <a class="btn btn-light btn-xl js-scroll-trigger" href="#services"
-              >Get Started!</a
-            >
-          </div>
-        </div>
-      </div>
-    </section>
-    <!-- Services-->
-    <section class="page-section" id="services">
-      <div class="container">
-        <h2 class="text-center mt-0">At Your Service</h2>
-        <hr class="divider my-4" />
-        <div class="row">
-          <div class="col-lg-3 col-md-6 text-center">
-            <div class="mt-5">
-              <i class="fas fa-4x fa-gem text-primary mb-4"></i>
-              <h3 class="h4 mb-2">Create Budget</h3>
-              <p class="text-muted mb-0">
-                Create your monthly budget
-              </p>
-            </div>
-          </div>
-          <div class="col-lg-3 col-md-6 text-center">
-            <div class="mt-5">
-              <i class="fas fa-4x fa-laptop-code text-primary mb-4"></i>
-              <h3 class="h4 mb-2">Track Expense</h3>
-              <p class="text-muted mb-0">
-                Track your expenses and categories
-              </p>
-            </div>
-          </div>
-          <div class="col-lg-3 col-md-6 text-center">
-            <div class="mt-5">
-              <i class="fas fa-4x fa-globe text-primary mb-4"></i>
-              <h3 class="h4 mb-2">Dashboard</h3>
-              <p class="text-muted mb-0">
-                Visualize your budget!
-              </p>
-            </div>
-          </div>
-        </div>
-      </div>
-    </section>
-    <!-- Footer-->
-    <footer class="bg-light py-5">
-      <div class="container">
-        <div class="small text-center text-muted">
-          Copyright © 2020 - Start Bootstrap
-        </div>
-      </div>
-    </footer>
-    <!-- Bootstrap core JS-->
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
-    <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.0/js/bootstrap.bundle.min.js"></script>
-    <!-- Third party plugin JS-->
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery-easing/1.4.1/jquery.easing.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/magnific-popup.js/1.1.0/jquery.magnific-popup.min.js"></script>
-    <!-- Core theme JS-->
-    <script src="/js/scripts.js"></script>
-  </body>
-</html>
